requirements.txt
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
blis==1.3.3
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contractions==0.1.73
cymem==2.0.13
defusedxml==0.7.1
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
funcy==2.0
gensim==4.4.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jpype1==1.6.0
kiwipiepy==0.22.0
kiwipiepy_model==0.22.0
konlpy==0.6.0
language_tool_python==2.9.5
load-dotenv==0.1.0
lxml==6.0.2
MarkupSafe==3.0.3
murmurhash==1.0.15
nltk==3.9.2
numexpr==2.14.1
numpy==2.3.5
openai==2.8.1
orjson==3.11.4
packaging==24.2
pandas==2.3.3
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
preshed==3.0.12
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
py-hanspell @ git+https://github.com/ssut/py-hanspell.git@fdc6ca50c19f1c85971437a072d89d4e5ce024b8
pyahocorasick==2.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pyLDAvis==3.4.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
textblob==0.19.0
textsearch==0.0.24
thinc==8.3.10
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
youtube-transcript-api==1.2.3
